fix: fix test setup imports
