+++ packages/graphback-runtime-mongodb/tests/MongoDataProviderTest.ts
@@ -1,7 +1,44 @@
+//tslint:disable-next-line: match-default-export-name
 import { MongoMemoryServer } from 'mongodb-memory-server';
-import { ObjectID } from 'mongodb';
+import { ObjectID, MongoClient } from 'mongodb';
+import { buildSchema } from 'graphql';
+import { filterModelTypes } from '@graphback/core';
 import { MongoDBDataProvider } from '../src/MongoDBDataProvider';
-import { createTestingContext, Context } from "./__util__";
+
+
+
+export interface Context {
+  providers: { [modelname: string]: MongoDBDataProvider };
+  server: MongoMemoryServer
+}
+
+export async function createTestingContext(schemaStr: string, config?: { seedData: { [collection: string]: any[] } }): Promise<Context> {
+    // Setup graphback
+    const schema = buildSchema(schemaStr);
+  
+    const server = new MongoMemoryServer();
+    const client = new MongoClient(await server.getConnectionString(), { useUnifiedTopology: true });
+    await client.connect();
+    const db = client.db('test');
+  
+    const providers: { [name: string]: MongoDBDataProvider } = {}
+    const models = filterModelTypes(schema)
+    for (const modelType of models) {
+      providers[modelType.name] = new MongoDBDataProvider(modelType, db);
+    }
+  
+    // if seed data is supplied, insert it into collections
+    if (config?.seedData) {
+      const collectionNames = Object.keys(config.seedData);
+      for (const collectionName of collectionNames) {
+        for (const element of config.seedData[collectionName]) {
+          await providers[collectionName].create(element);
+        }
+      };
+    }
+  
+    return { server, providers }
+  }
 
 describe('MongoDBDataProvider Basic CRUD', () => {
   interface Todo {
